- Fixed: Thing_SetTranslation still used a 16-bit word to hold the translation. - Bumped the maximum resolution up to 2560x1600. - Fixed: DCanvas::DrawTexture() only expanded virtual screen sizes for widescreen   resolutions but left 5:4 modes alone. This fix neccessitated the addition of   DTA_Bottom320x200 for the status bar to use so that it could be flush with the   bottom of the screen rather than sitting slightly above it. - Fixed: FConfigFile::ReadConfig()'s definition of whitespace was too broad. - Fixed: Defining custom translation ranges in descending order and/or with gaps   made for crashes.
SVN r676 (trunk)

+++ src/p_lnspec.cpp
@@ -1496,7 +1496,7 @@ FUNC(LS_Thing_SetTranslation)
 // Thing_SetTranslation (tid, range)
 {
 	TActorIterator<AActor> iterator (arg0);
-	WORD range;
+	int range;
 	AActor *target;
 	bool ok = false;
